clang-tidy: fix bugprone-narrowing-conversions

+++ src/tss2-fapi/fapi_crypto.c
@@ -752,7 +752,7 @@ ifapi_ecc_der_sig_to_tpm(
     const BIGNUM *bnr;
     const BIGNUM *bns;
 
-    d2i_ECDSA_SIG(&ecdsaSignature, &signature, signatureSize);
+    d2i_ECDSA_SIG(&ecdsaSignature, &signature, (long) signatureSize);
     return_if_null(ecdsaSignature, "Invalid DER signature",
                    TSS2_FAPI_RC_GENERAL_FAILURE);
 
@@ -1188,12 +1188,12 @@ get_ecc_tpm2b_public_from_evp(
     tpmPublic->publicArea.unique.ecc.x.size = ecKeySize;
     tpmPublic->publicArea.unique.ecc.y.size = ecKeySize;
     if (1 != ifapi_bn2binpad(bnX, &tpmPublic->publicArea.unique.ecc.x.buffer[0],
-                             ecKeySize)) {
+                             (int) ecKeySize)) {
         goto_error(r, TSS2_FAPI_RC_GENERAL_FAILURE,
                    "Write big num byte buffer", cleanup);
     }
     if (1 != ifapi_bn2binpad(bnY, &tpmPublic->publicArea.unique.ecc.y.buffer[0],
-                             ecKeySize)) {
+                             (int) ecKeySize)) {
         goto_error(r, TSS2_FAPI_RC_GENERAL_FAILURE,
                    "Write big num byte buffer", cleanup);
     }
@@ -1253,7 +1253,7 @@ ifapi_get_evp_from_pem(const char *pemKey, EVP_PKEY **publicKey) {
     BIO *bufio = NULL;
 
     /* Use BIO for conversion */
-    bufio = BIO_new_mem_buf((void *)pemKey, strlen(pemKey));
+    bufio = BIO_new_mem_buf((void *)pemKey, (int) strlen(pemKey));
     goto_if_null(bufio, "BIO buffer could not be allocated.",
                  TSS2_FAPI_RC_MEMORY, cleanup);
 
@@ -1418,7 +1418,7 @@ ifapi_verify_signature_quote(
 
     /* Create an OpenSSL object for the key */
     bufio = BIO_new_mem_buf((void *)public_pem_key,
-                            strlen(public_pem_key));
+                            (int) strlen(public_pem_key));
     goto_if_null(bufio, "BIO buffer could not be allocated.",
                  TSS2_FAPI_RC_MEMORY, error_cleanup);
 
@@ -1549,7 +1549,7 @@ ifapi_verify_signature(
 
     /* Convert the key to an OpenSSL object */
     bufio = BIO_new_mem_buf((void *)public_pem_key,
-                                strlen(public_pem_key));
+                                (int) strlen(public_pem_key));
     goto_if_null(bufio, "Out of memory.", TSS2_FAPI_RC_MEMORY, error_cleanup);
     publicKey = PEM_read_bio_PUBKEY(bufio, NULL, NULL, NULL);
     goto_if_null(publicKey, "PEM format could not be decoded.",
@@ -1823,7 +1823,7 @@ ifapi_cert_to_pem(
     EVP_PKEY *publicKey = NULL;
     int pemCertSize;
 
-    if (!d2i_X509(&cert, (const unsigned char **)&certBuffer, certBufferSize)) {
+    if (!d2i_X509(&cert, (const unsigned char **)&certBuffer, (long) certBufferSize)) {
         LOGBLOB_ERROR(certBuffer, certBufferSize, "Bad certificate data");
         return_error(TSS2_FAPI_RC_GENERAL_FAILURE, "Invalid certificate.");
     }
@@ -1939,7 +1939,7 @@ static X509
 
     /* Use BIO for conversion */
     size_t pem_length = strlen(pem_cert);
-    bufio = BIO_new_mem_buf((void *)pem_cert, pem_length);
+    bufio = BIO_new_mem_buf((void *)pem_cert, (int) pem_length);
     if (!bufio)
         return NULL;
     /* Convert the certificate */
@@ -2096,7 +2096,7 @@ ifapi_base64encode(uint8_t *buffer, size_t buffer_size, char** b64_data) {
     bio = BIO_push(bio64, bio);
 
     BIO_set_flags(bio, BIO_FLAGS_BASE64_NO_NL);
-    bytes_written = BIO_write(bio, buffer, buffer_size);
+    bytes_written = BIO_write(bio, buffer, (int) buffer_size);
     if (bytes_written != (int)buffer_size) {
         goto_error(r, TSS2_FAPI_RC_GENERAL_FAILURE, "Invalid BIO_write",
                    cleanup);
@@ -2149,7 +2149,7 @@ ifapi_rsa_encrypt(const char *pem_key,
     const EVP_MD *evp_md;
 
     /* Convert the pem key to an OpenSSL object */
-    bufio = BIO_new_mem_buf((void *)pem_key, strlen(pem_key));
+    bufio = BIO_new_mem_buf((void *)pem_key, (int) strlen(pem_key));
     goto_if_null(bufio, "Out of memory.", TSS2_FAPI_RC_MEMORY, cleanup);
 
     publicKey = PEM_read_bio_PUBKEY(bufio, NULL, NULL, NULL);
@@ -2332,7 +2332,7 @@ load_private_ECC_from_key(EVP_PKEY *key,
         goto out;
     }
 
-    p->size = BN_bn2binpad(b, p->buffer, priv_bytes);
+    p->size = BN_bn2binpad(b, p->buffer, (int) priv_bytes);
     if (p->size != priv_bytes) {
         goto out;
     }
