+++ src/tss2-tcti/tcti-i2c-helper.h
@@ -36,7 +36,7 @@ typedef struct {
     TSS2_TCTI_I2C_HELPER_PLATFORM platform;
     bool guard_time_read;
     bool guard_time_write;
-    int  guard_time;
+    uint8_t  guard_time;
     char header[TCTI_I2C_HELPER_RESP_HEADER_SIZE];
 } TSS2_TCTI_I2C_HELPER_CONTEXT;
